MW-SO_1.0.9_Sales_Return_updated

+++ MWSO/app/src/main/java/com/techsmith/mw_so/retail_utils/PAYSUMMARY.java
@@ -1,6 +1,6 @@
 package com.techsmith.mw_so.retail_utils;
 
 public class PAYSUMMARY {
-    public float BILLAMOUNT;
-    public float PAIDAMOUNT;
+    public double BILLAMOUNT;
+    public double PAIDAMOUNT;
 }
